display optimize change
