+++ include/commands/diff.hpp
@@ -1,0 +1,33 @@
+#ifndef DIFF_HPP
+#define DIFF_HPP
+
+#include <string>
+#include <vector>
+#include "../utils/kit_utils.hpp"
+
+namespace kit_vcs
+{
+    inline std::vector<std::string> get_differences(const std::string &commit_hash)
+    {
+        std::vector<std::string> differences;
+
+        if (!kit_utils::ensure_repository_initialized())
+        {
+            return differences;
+        }
+
+        try
+        {
+            // Placeholder logic for computing differences
+            differences.push_back("Difference computation is not yet implemented for commit: " + commit_hash);
+        }
+        catch (const std::exception &e)
+        {
+            kit_utils::print_error("Failed to get differences: " + std::string(e.what()));
+        }
+
+        return differences;
+    }
+} // namespace kit_vcs
+
+#endif // DIFF_HPP
